+++ tests/msw/handlers.server.ts
@@ -2,6 +2,7 @@ import { clientIdHandlers } from "./domains/client-ids.handlers";
 import { currenciesHandlers } from "./domains/currencies.handlers";
 import { miscHandlers } from "./domains/misc.handlers";
 import { payHandlers } from "./domains/pay.handlers";
+import { payeeDestinationHandlers } from "./domains/payee-destination.handlers";
 import { payerHandlers } from "./domains/payer.handlers";
 import { paymentsHandlers } from "./domains/payments.handlers";
 import { payoutsHandlers } from "./domains/payouts.handlers";
@@ -11,6 +12,7 @@ export const handlers = [
   ...currenciesHandlers,
   ...requestsHandlers,
   ...payHandlers,
+  ...payeeDestinationHandlers,
   ...payerHandlers,
   ...paymentsHandlers,
   ...payoutsHandlers,